pickem/requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
autopep8==2.0.2
beautifulsoup4==4.11.1
boto3==1.24.65
botocore==1.27.65
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.51.0
django-apscheduler==0.6.2
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-storages==1.13.1
django-tables2==2.4.1
djangorestframework==3.13.1
espn-api==0.45.1
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
